golang-tools: delay the upcoming call
I had completely missed that later this week I have a conflict
with KubeCon EU, so I won't be able to attend nor run it.
Moreover, multiple Googlers won't be around this Wednesday either.

I'll be travelling the two weeks after KubeCon, so push it back
a bit further as well, unfortunately. At least the agenda was empty
for now, so it seems like there wasn't much to go through this week.

Change-Id: I9c2a8276f1a10c5167c8b13fd70561d40c11b330
Reviewed-on: https://go-review.googlesource.com/c/wiki/+/572395
Reviewed-by: Robert Findley <rfindley@google.com>