add "cue" as a well-known struct tag
As described in the linked Go package documentation page,
the struct tag is used to validate existing Go types
with CUE expressions such as constraints.

It's hard to say what qualifies as "well known",
but there are previous examples in the list which don't seem
very widely used, and a quick search on GitHub via

    https://github.com/search?q=%22%60cue%3A%22+language%3Ago&type=code

shows a hundred matches today.

Change-Id: Ie8cd6650bfc9d869e4407cda5694839da13c57e3
Reviewed-on: https://go-review.googlesource.com/c/wiki/+/587895
Reviewed-by: Ingo Oeser <nightlyone@googlemail.com>
Reviewed-by: Ian Lance Taylor <iant@golang.org>