DashboardBuilders: remove steps to add builders to legacy system
The migration to LUCI is far along now that we aren't adding any more
new builders to coordinator. Remove the instructions for doing that,
and also move the LUCI section to be on top. This will unfortunately
require some links to be updated, but there's no way to avoid that.

Keep the security note section since its content generally applies
to the new system too.

For golang/go#67761.

Change-Id: I338b5622b1f9020539d1ab309f39127a4f3e0c93
Reviewed-on: https://go-review.googlesource.com/c/wiki/+/592755
Reviewed-by: Carlos Amedee <carlos@golang.org>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>