CreatingSubRepository: update for migration to LUCI
Include a few changes that are relevant now that LUCI is used for
testing. Notably, the luci-config branch needs to be updated, and
it's fine not to redeploy coordinator. Clarify that the x/website
deployments are automatic, and add a reminder to add repo owner(s).

Change-Id: Idc49c555cc527c23aa348bc42597064c78e52d25
Reviewed-on: https://go-review.googlesource.com/c/wiki/+/596595
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Than McIntosh <thanm@google.com>