wiki: Prefer "white space" over "whitespace"
Both "whitespace" and "white space" are used in the codebase.
There doesn't seem to be an external consensus as to which is better
(see https://github.com/orgs/mdn/discussions/382 for a discussion).
https://www.unicode.org/versions/Unicode16.0.0/UnicodeStandard-16.0.pdf
(warning: 12 MB file) consistently uses "whitespace", but the character
property is "White_Space".

Regardless of which, if any, term is deemed to be preferred, I
believe the Go codebase should at least be consistent.

This change was prompted by browsing pkg.go.dev/strings and
searching for "whitespace", which found only strings.FieldsSeq,
not strings.Fields, which uses "white space".

Change-Id: I0ad850831e4359a8c84afffa4ad86373e14133ca
Reviewed-on: https://go-review.googlesource.com/c/wiki/+/656315
Reviewed-by: Rob Pike <r@golang.org>
Reviewed-by: Ian Lance Taylor <iant@golang.org>
Auto-Submit: Ian Lance Taylor <iant@golang.org>