GerritBot.md: add summary bullets and image for editing commit message text in GitHub
The FAQ on "How does GerritBot determine the final commit message" is a
link destination from various documentation and automatic messages from
GerritBot. Add some summary bullets at the start along with a simple
image showing where to edit in the GitHub PR.

The FAQ on "I left a reply to a comment in Gerrit but no one but me can
see it" is similarly a link destination. Add a reminder there that the
user should be logged in to Gerrit, especially if they don't see the
"Reply" button.

Change-Id: I17bab63acddbbe328d08ababfebaceb37ea8e0f3
Reviewed-on: https://go-review.googlesource.com/c/wiki/+/691076
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>